requirements.txt
flask
pyodbc
pandas
matplotlib
reportlab
python-dotenv
gunicorn
